Silent update walkthrough
